// frontend/public/game/image_loader.d.ts
export declare class ImageSrc {
    playerImg: HTMLImageElement | HTMLCanvasElement;
    ai_target: HTMLImageElement | HTMLCanvasElement;
    powerupImg: HTMLImageElement | HTMLCanvasElement;
    obstacleImg: HTMLImageElement | HTMLCanvasElement;
    meteorImg: HTMLImageElement | HTMLCanvasElement;
    top_img: HTMLImageElement | HTMLCanvasElement;
    start_img: HTMLImageElement | HTMLCanvasElement;
    startcustom_img: HTMLImageElement | HTMLCanvasElement;
    font0: HTMLImageElement;
    font1: HTMLImageElement;
    font2: HTMLImageElement;
    font3: HTMLImageElement;
    font4: HTMLImageElement;
    font5: HTMLImageElement;
    font6: HTMLImageElement;
    font7: HTMLImageElement;
    font8: HTMLImageElement;
    font9: HTMLImageElement;
    nbrfont: HTMLImageElement[];
    constructor();
    reloadColorImage(main_color: string): void;
}
export declare function hexToRgbArray(hex: string): number[];
//# sourceMappingURL=image_loader.d.ts.map